refactor extract CustomTab
